refactor: split script.js into dom.js, chat.js, and video.js for better modularity
- moved all DOM manipulation and event binding logic into dom.js
- moved socket-based chat functionality into chat.js
- moved peer-to-peer video streaming logic into video.js
- simplified script.js to act as the main entry point coordinating modules
