+++ public/index.html
@@ -12,6 +12,7 @@
     />
     <link rel="stylesheet" href="css/styles.css" />
     <title>Omegle Demo : Talk to Strangers !</title>
+    <script type="module" src="js/script.js"></script>
   </head>
   <body>
     <div class="join-container">
@@ -117,6 +118,5 @@
     ></script>
     <script src="https://unpkg.com/peerjs@1.5.4/dist/peerjs.min.js"></script>
     <script src="/socket.io/socket.io.js"></script>
-    <script src="js/script.js"></script>
   </body>
 </html>
